tests/wpt/tests/html/dom/render-blocking/element-render-blocking-022.tentative.html

<!DOCTYPE html>
<head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/utils.js"></script>
<title>Href attribute changes in the body to apply, but has no effect</title>

<link id=link rel=expect href="" blocking="render">
<script>
async_test((t) => {
  requestAnimationFrame(() => {
    t.step(() => assert_false(!!document.getElementById("last")));
    t.done();
  });
}, "adding href in the body has no effect");
</script>
</head>
<body>
<script>
link.href = "#last";
</script>

  <div id="first"></div>
  <script>
    jankMany(100, 10);
  </script>
  <div id="second"></div>
  <script>
    jankMany(100, 10);
  </script>
  <div id="last"></div>
</body>
